refactor(rc-1): extract remaining 15 tools to lib/tools/*
Completes RC-1 tool extraction started in PR #115. Each of the 15 remaining inline codex-* tool definitions in index.ts is moved to its own factory module under lib/tools/, following the ToolContext pattern documented in lib/tools/AGENTS.md.

- 15 new lib/tools/codex-*.ts factory files (codex-list, codex-switch, codex-status, codex-limits, codex-metrics, codex-doctor, codex-label, codex-tag, codex-note, codex-dashboard, codex-health, codex-remove, codex-refresh, codex-export, codex-import)

- lib/tools/index.ts: createToolRegistry wires all 18 factories; ToolContext.buildRoutingVisibilitySnapshot selectionExplainability param uses AccountSelectionExplainability[] instead of unknown

- index.ts: inline tool defs replaced with single createToolRegistry(ctx) call; cachedAccountManager/accountManagerPromise threaded via getter/setter MutableRefs so tool writes propagate. Line count 6269 -> 3425 (-45%)

- lib/tools/AGENTS.md: status updated from 'incremental follow-up pending' to 'complete'

Behavior unchanged. All 2121 tests pass. typecheck, lint, build all green.
